requirements.txt
streamlit
pandas
scikit-learn
joblib
newspaper3k
lxml_html_clean
xgboost